Split program.js into expression, primitive, and program
The original program.js is way too long and mixing different ast nodes all in one place.
This split should make more sense.
