+++ lib/ast/bookkeeping.js
@@ -11,8 +11,9 @@
 
 const adt = require('adt');
 
-const { Input, BooleanExpression } = require('./program');
+const { Input } = require('./program');
 const { Value } = require('./values');
+const { BooleanExpression } = require('./expression');
 
 /**
  * A ThingTalk input that drives the dialog.
