+++ lib/ast/class_def.js
@@ -15,7 +15,8 @@ const Type = require('../type');
 const { prettyprintClassDef } = require('../prettyprint');
 const { clean } = require('../utils');
 const { Value } = require('./values');
-const { Statement, InputParam } = require('./program');
+const { InputParam } = require('./primitive');
+const { Statement } = require('./program');
 const { FunctionDef } = require('./function_def');
 const { getString, extractImports, typeToHTML } = require('./manifest_utils');
 const toJS = require('./toJS');
