+++ lib/ast/index.js
@@ -15,6 +15,8 @@ const FunctionDefs = require('./function_def');
 const ClassDefs = require('./class_def');
 const ComputeDefs = require('./compute_def');
 const Values = require('./values');
+const Expressions = require('./expression');
+const Primitives = require('./primitive');
 const Programs = require('./program');
 const Bookkeepings = require('./bookkeeping');
 
@@ -40,5 +42,5 @@ adt.nativeClone = function nativeClone(x) {
  *
  * @namespace Ast
  */
-Object.assign(module.exports, FunctionDefs, ClassDefs, ComputeDefs, Values, Programs, Bookkeepings);
+Object.assign(module.exports, FunctionDefs, ClassDefs, ComputeDefs, Values, Expressions, Primitives, Programs, Bookkeepings);
 require('./api');
