+++ lib/ast/manifest_utils.js
@@ -11,7 +11,8 @@
 
 const Type = require('../type');
 const { Value } = require('./values');
-const { InputParam, ImportStmt } = require('./program');
+const { InputParam } = require('./primitive');
+const { ImportStmt } = require('./program');
 
 function getString(value, fallback='') {
     if (!value) return fallback;
