+++ lib/ast/program.js
@@ -13,1359 +13,11 @@ const assert = require('assert');
 
 const Type = require('../type');
 const Base = require('./base');
-const { ExpressionSignature } = require('./function_def');
 const { Value } = require('./values');
+const { BooleanExpression } = require('./expression');
+const { Stream, Table, Action, PermissionFunction } = require('./primitive');
 const toJS = require('./toJS');
 const { recursiveYieldArraySlots, FieldSlot } = require('./slots');
-
-/**
- * Base class of all expressions that select a device.
- *
- * Selectors correspond to the `@`-device part of the ThingTalk code,
- * up to but not including the function name.
- *
- * @alias Ast.Selector
- * @extends Ast.Base
- * @property {boolean} isSelector - true
- * @property {boolean} isDevice - true if this is an instance of {@link Ast.Selector.Device}
- * @property {boolean} isBuiltin - true if this is {@link Ast.Selector.Builtin}
- * @abstract
- */
-class Selector extends Base {}
-Selector.prototype.isSelector = true;
-/**
- * A selector that maps to one or more devices in Thingpedia.
- *
- * @alias Ast.Selector.Device
- * @extends Ast.Selector
- * @param {string} kind - the Thingpedia class ID
- * @param {string|null} id - the unique ID of the device being selected, or null
- *                           to select all devices
- * @param {null} principal - reserved/deprecated, must be `null`
- */
-class DeviceSelector extends Selector {
-    constructor(kind, id, principal) {
-        super();
-
-        assert(typeof kind === 'string');
-        this.kind = kind;
-
-        assert(typeof id === 'string' || id === null);
-        this.id = id;
-
-        assert(principal === null);
-        this.principal = principal;
-    }
-
-    clone() {
-        return new DeviceSelector(this.kind, this.id, this.principal);
-    }
-
-    toString() {
-        return `Device(${this.kind}, ${this.id ? this.id : ''}, ${this.principal ? this.principal : ''})`;
-    }
-}
-DeviceSelector.prototype.isDevice = true;
-Selector.Device = DeviceSelector;
-/**
- * A selector that maps the builtin `notify`, `return` and `save` functions.
- *
- * @alias Ast.Selector.Builtin
- * @extends Ast.Selector
- * @readonly
- */
-class BuiltinDevice extends Selector {
-    clone() {
-        return new BuiltinDevice();
-    }
-
-    toString() {
-        return 'Builtin';
-    }
-}
-BuiltinDevice.prototype.isBuiltin = true;
-Selector.Builtin = new BuiltinDevice();
-module.exports.Selector = Selector;
-
-
-/**
- * An invocation of a ThingTalk function.
- *
- * @alias Ast.Invocation
- * @extends Ast.Base
- * @param {Ast.Selector} selector - the selector choosing where the function is invoked
- * @param {string} channel - the function name
- * @param {Ast.InputParam[]} in_params - input parameters passed to the function
- * @param {Ast.ExpressionSignature|null} schema - type signature of the invoked function
- * @property {boolean} isInvocation - true
- */
-class Invocation extends Base {
-    constructor(selector, channel, in_params, schema) {
-        super();
-
-        assert(selector instanceof Selector);
-        this.selector = selector;
-
-        assert(typeof channel === 'string');
-        this.channel = channel;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-
-    clone() {
-        return new Invocation(
-            this.selector.clone(),
-            this.channel,
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone(): null
-        );
-    }
-
-    toString() {
-        const in_params = this.in_params && this.in_params.length > 0 ? this.in_params.toString() : '';
-        return `Invocation(${this.selector.toString()}, ${this.channel}, ${in_params}, )`;
-    }
-}
-Invocation.prototype.isInvocation = true;
-module.exports.Invocation = Invocation;
-
-
-/**
- * An expression that computes a boolean predicate.
- * This AST node is used in filter expressions.
- *
- * @class
- * @alias Ast.BooleanExpression
- * @extends Ast.Base
- * @abstract
- * @property {boolean} isBooleanExpression - true
- * @property {boolean} isAnd - true if this is an instance of {@link Ast.BooleanExpression.And}
- * @property {boolean} isOr - true if this is an instance of {@link Ast.BooleanExpression.Or}
- * @property {boolean} isAtom - true if this is an instance of {@link Ast.BooleanExpression.Atom}
- * @property {boolean} isNot - true if this is an instance of {@link Ast.BooleanExpression.Not}
- * @property {boolean} isExternal - true if this is an instance of {@link Ast.BooleanExpression.External}
- * @property {boolean} isTrue - true if this is {@link Ast.BooleanExpression.True}
- * @property {boolean} isFalse - true if this is {@link Ast.BooleanExpression.False}
- * @property {boolean} isVarRef - true if this is {@link Ast.BooleanExpression.VarRef}
- * @property {boolean} isCompute - true if this is {@link Ast.BooleanExpression.Compute}
- */
-class BooleanExpression extends Base {}
-BooleanExpression.prototype.isBooleanExpression = true;
-/**
- * A conjunction boolean expression (ThingTalk operator `&&`)
- * @alias Ast.BooleanExpression.And
- * @extends Ast.BooleanExpression
- * @param {Ast.BooleanExpression[]} operands - the expression operands
- */
-class AndBooleanExpression extends BooleanExpression {
-    constructor(operands) {
-        super();
-
-        assert(Array.isArray(operands));
-        this.operands = operands;
-    }
-
-    clone() {
-        return new AndBooleanExpression(
-            this.operands.map((operand) => operand.clone())
-        );
-    }
-}
-BooleanExpression.And = AndBooleanExpression;
-BooleanExpression.And.prototype.isAnd = true;
-/**
- * A disjunction boolean expression (ThingTalk operator `||`)
- * @alias Ast.BooleanExpression.Or
- * @extends Ast.BooleanExpression
- * @param {Ast.BooleanExpression[]} operands - the expression operands
- */
-class OrBooleanExpression extends BooleanExpression {
-    constructor(operands) {
-        super();
-
-        assert(Array.isArray(operands));
-        this.operands = operands;
-    }
-
-    clone() {
-        return new OrBooleanExpression(
-            this.operands.map((operand) => operand.clone())
-        );
-    }
-}
-BooleanExpression.Or = OrBooleanExpression;
-BooleanExpression.Or.prototype.isOr = true;
-/**
- * A comparison expression (predicate atom)
- * @alias Ast.BooleanExpression.Atom
- * @extends Ast.BooleanExpression
- * @param {string} name - the parameter name to compare
- * @param {string} operator - the comparison operator
- * @param {Ast.Value} value - the value being compared against
- */
-class AtomBooleanExpression extends BooleanExpression {
-    constructor(name, operator, value) {
-        super();
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(typeof operator === 'string');
-        this.operator = operator;
-
-        assert(value instanceof Value);
-        this.value = value;
-    }
-
-    clone() {
-        return new AtomBooleanExpression(
-            this.name, this.operator, this.value.clone()
-        );
-    }
-
-    toString() {
-        return `Atom(${this.name}, ${this.operator}, ${this.value})`;
-    }
-}
-BooleanExpression.Atom = AtomBooleanExpression;
-BooleanExpression.Atom.prototype.isAtom = true;
-/**
- * A negation boolean expression (ThingTalk operator `!`)
- * @alias Ast.BooleanExpression.Not
- * @extends Ast.BooleanExpression
- * @param {Ast.BooleanExpression} expr - the expression being negated
- */
-class NotBooleanExpression extends BooleanExpression {
-    constructor(expr) {
-        super();
-
-        assert(expr instanceof BooleanExpression);
-        this.expr = expr;
-    }
-
-    clone() {
-        return new NotBooleanExpression(this.expr.clone());
-    }
-}
-BooleanExpression.Not = NotBooleanExpression;
-BooleanExpression.Not.prototype.isNot = true;
-/**
- * A boolean expression that calls a Thingpedia query function
- * and filters the result.
- *
- * The boolean expression is true if at least one result from the function
- * call satisfies the filter.
- *
- * @alias Ast.BooleanExpression.External
- * @extends Ast.BooleanExpression
- * @param {Ast.Selector.Device} selector - the selector choosing where the function is invoked
- * @param {string} channel - the function name
- * @param {Ast.InputParam[]} in_params - input parameters passed to the function
- * @param {Ast.BooleanExpression} filter - the filter to apply on the invocation's results
- * @param {Ast.ExpressionSignature|null} schema - type signature of the invoked function
- */
-class ExternalBooleanExpression extends BooleanExpression {
-    constructor(selector, channel, in_params, filter, schema) {
-        super();
-
-        assert(selector instanceof Selector);
-        this.selector = selector;
-
-        assert(typeof channel === 'string');
-        this.channel = channel;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-
-        assert(filter instanceof BooleanExpression);
-        this.filter = filter;
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-
-    clone() {
-        return new ExternalBooleanExpression(
-            this.selector.clone(),
-            this.channel,
-            this.in_params.map((p) => p.clone()),
-            this.filter.clone(),
-            this.schema ? this.schema.clone(): null
-        );
-    }
-}
-BooleanExpression.External = ExternalBooleanExpression;
-BooleanExpression.External.prototype.isExternal = true;
-/**
- * The constant `true` boolean expression.
- *
- * @alias Ast.BooleanExpression.True
- * @extends Ast.BooleanExpression
- * @readonly
- */
-class TrueBooleanExpression extends BooleanExpression {
-    clone() {
-        return new TrueBooleanExpression();
-    }
-}
-TrueBooleanExpression.prototype.isTrue = true;
-BooleanExpression.True = new TrueBooleanExpression();
-/**
- * The constant `false` boolean expression.
- *
- * @alias Ast.BooleanExpression.False
- * @extends Ast.BooleanExpression
- * @readonly
- */
-class FalseBooleanExpression extends BooleanExpression {
-    clone() {
-        return new FalseBooleanExpression();
-    }
-}
-FalseBooleanExpression.prototype.isFalse = true;
-BooleanExpression.False = new FalseBooleanExpression();
-/**
- * A boolean expression that calls a Thingpedia computation macro
- *
- *
- * @alias Ast.BooleanExpression.VarRef
- * @extends Ast.BooleanExpression
- * @param {Ast.Selector.Device} selector - the selector choosing where the function is invoked
- * @param {string} name - the macro name
- * @param {Ast.InputParam[]} in_params - input parameters passed to the macro
- */
-class VarRefBooleanExpression extends BooleanExpression {
-    constructor(selector, name, args) {
-        super();
-
-        assert(selector instanceof Selector);
-        this.selector = selector;
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(Array.isArray(args));
-        this.args = args;
-    }
-
-    clone() {
-        return new VarRefBooleanExpression(
-            this.selector.clone(),
-            this.name,
-            this.args.map((a) => a)
-        );
-    }
-}
-BooleanExpression.VarRef = VarRefBooleanExpression;
-BooleanExpression.VarRef.prototype.isVarRef = true;
-/**
- * A boolean expression that computes a scalar expression and then do a comparison
- *
- *
- * @alias Ast.BooleanExpression.Compute
- * @extends Ast.BooleanExpression
- * @param {Ast.ScalarExpression} lhs - the scalar expression to compute
- * @param {string} operator - the comparison operator
- * @param {Ast.Value} value - the value being compared against
- */
-class ComputeBooleanExpression extends BooleanExpression {
-    constructor(lhs, operator, rhs) {
-        super();
-
-        assert(lhs instanceof ScalarExpression);
-        this.lhs = lhs;
-
-        assert(typeof operator === 'string');
-        this.operator = operator;
-
-        assert(rhs instanceof Value);
-        this.rhs = rhs;
-    }
-
-    clone() {
-        return new ComputeBooleanExpression(
-            this.lhs.clone(),
-            this.operator,
-            this.rhs.clone()
-        );
-    }
-}
-BooleanExpression.Compute = ComputeBooleanExpression;
-BooleanExpression.Compute.prototype.isCompute = true;
-module.exports.BooleanExpression = BooleanExpression;
-
-
-class ListExpression extends Base {
-    constructor(name, filter) {
-        super();
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(filter === null || filter instanceof BooleanExpression);
-        this.filter = filter;
-    }
-
-    clone() {
-        return new ListExpression(
-            this.name,
-            this.filter ? this.filter.clone() : null
-        );
-    }
-}
-ListExpression.prototype.isListExpression = true;
-module.exports.ListExpression = ListExpression;
-
-
-class ScalarExpression extends Base {}
-ScalarExpression.prototype.isScalarExpression = true;
-class PrimaryScalarExpression extends ScalarExpression {
-    constructor(value) {
-        super();
-
-        assert(value instanceof Value);
-        this.value = value;
-    }
-
-    clone() {
-        return new PrimaryScalarExpression(this.value.clone());
-    }
-}
-ScalarExpression.Primary = PrimaryScalarExpression;
-ScalarExpression.Primary.prototype.isPrimary = true;
-class DerivedScalarExpression extends ScalarExpression {
-    constructor(op, operands) {
-        super();
-
-        assert(typeof op === 'string');
-        this.op = op;
-
-        assert(Array.isArray(operands));
-        this.operands = operands;
-    }
-
-    clone() {
-        return new DerivedScalarExpression(this.op, this.operands.map((operand) => operand.clone()));
-    }
-}
-ScalarExpression.Derived = DerivedScalarExpression;
-ScalarExpression.Derived.prototype.isDerived = true;
-class AggregationScalarExpression extends ScalarExpression {
-    constructor(operator, field, list) {
-        super();
-
-        assert(typeof operator === 'string');
-        this.operator = operator;
-
-        assert(field === null || typeof field === 'string');
-        this.field = field;
-
-        assert(list instanceof ListExpression);
-        this.list = list;
-    }
-
-    clone() {
-        return new AggregationScalarExpression(
-            this.operator,
-            this.field,
-            this.list.clone()
-        );
-    }
-}
-ScalarExpression.Aggregation = AggregationScalarExpression;
-ScalarExpression.Aggregation.prototype.isAggregation = true;
-class FilterScalarExpression extends ScalarExpression {
-    constructor(list) {
-        super();
-
-        assert(list instanceof ListExpression);
-        this.list = list;
-    }
-
-    clone() {
-        return new FilterScalarExpression(this.list.clone());
-    }
-}
-ScalarExpression.Filter = FilterScalarExpression;
-ScalarExpression.Filter.prototype.isFilter = true;
-class FlattenedListScalarExpression extends ScalarExpression {
-    constructor(list) {
-        super();
-
-        assert(list instanceof ListExpression);
-        this.list = list;
-    }
-
-    clone() {
-        return new FlattenedListScalarExpression(this.list.clone());
-    }
-}
-ScalarExpression.FlattenedList = FlattenedListScalarExpression;
-ScalarExpression.FlattenedList.prototype.isFlattenedList = true;
-class VarRefScalarExpression extends ScalarExpression {
-    constructor(selector, name, args) {
-        super();
-
-        assert(selector instanceof Selector);
-        this.selector = selector;
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(Array.isArray(args));
-        this.args = args;
-    }
-
-    clone() {
-        return new VarRefScalarExpression(
-            this.selector.clone(),
-            this.name,
-            this.args.map((a) => a)
-        );
-    }
-}
-ScalarExpression.VarRef = VarRefScalarExpression;
-ScalarExpression.VarRef.prototype.isVarRef = true;
-module.exports.ScalarExpression = ScalarExpression;
-
-
-/**
- * AST node corresponding to an input parameter passed to a function.
- *
- * @class
- * @alias Ast.InputParam
- * @extends Ast.Base
- * @param {string} name - the input argument name
- * @param {Ast.Value} value - the value being passed
- * @property {boolean} isInputParam - true
- */
-class InputParam extends Base {
-    constructor(name, value) {
-        super();
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(value instanceof Value);
-        this.value = value;
-    }
-
-    clone() {
-        return new InputParam(this.name, this.value.clone());
-    }
-
-    toString() {
-        return `InputParam(${this.name}, ${this.value})`;
-    }
-}
-InputParam.prototype.isInputParam = true;
-module.exports.InputParam = InputParam;
-
-
-/**
- * The base class of all ThingTalk query expressions.
- *
- * @alias Ast.Table
- * @extends Ast.Base
- * @abstract
- * @param {Ast.ExpressionSignature|null} schema - type signature of the invoked function
- */
-class Table extends Base {
-    constructor(schema) {
-        super();
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-}
-Table.prototype.isTable = true;
-class VarRefTable extends Table {
-    constructor(name, in_params, schema) {
-        super(schema);
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-    }
-
-    clone() {
-        return new VarRefTable(
-            this.name,
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.VarRef = VarRefTable;
-Table.VarRef.prototype.isVarRef = true;
-class ResultRefTable extends Table {
-    constructor(kind, channel, index, schema) {
-        super(schema);
-
-        assert(typeof kind === 'string');
-        this.kind = kind;
-
-        assert(typeof channel === 'string');
-        this.channel = channel;
-
-        assert(index instanceof Value);
-        this.index = index;
-    }
-
-    clone() {
-        return new ResultRefTable(
-            this.kind,
-            this.channel,
-            this.index.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-
-    toString() {
-        return `ResultRef(${this.kind}, ${this.channel}, ${this.index}, )`;
-    }
-}
-Table.ResultRef = ResultRefTable;
-Table.ResultRef.prototype.isResultRef = true;
-class TableInvocation extends Table {
-    constructor(invocation, schema) {
-        super(schema);
-
-        assert(invocation instanceof Invocation);
-        this.invocation = invocation;
-    }
-
-    clone() {
-        return new TableInvocation(
-            this.invocation.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Invocation = TableInvocation;
-Table.Invocation.prototype.isInvocation = true;
-class FilteredTable extends Table {
-    constructor(table, filter, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(filter instanceof BooleanExpression);
-        this.filter = filter;
-    }
-
-    clone() {
-        return new FilteredTable(
-            this.table.clone(),
-            this.filter.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Filter = FilteredTable;
-Table.Filter.prototype.isFilter = true;
-class TableProjection extends Table {
-    constructor(table, args, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(Array.isArray(args));
-        this.args = args;
-    }
-
-    clone() {
-        return new TableProjection(
-            this.table.clone(),
-            this.args.map((a) => (a)),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Projection = TableProjection;
-Table.Projection.prototype.isProjection = true;
-class ComputeTable extends Table {
-    constructor(table, expression, alias, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(expression instanceof ScalarExpression);
-        this.expression = expression;
-
-        assert(alias === null || typeof alias === 'string');
-        this.alias = alias;
-    }
-
-    clone() {
-        return new ComputeTable(
-            this.table.clone(),
-            this.expression.clone(),
-            this.alias,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Compute = ComputeTable;
-Table.Compute.prototype.isCompute = true;
-class AliasTable extends Table {
-    constructor(table, name, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(typeof name === 'string');
-        this.name = name;
-    }
-
-    clone() {
-        return new AliasTable(
-            this.table.clone(),
-            this.name,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Alias = AliasTable;
-Table.Alias.prototype.isAlias = true;
-class AggregationTable extends Table {
-    constructor(table, field, operator, alias, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(typeof field === 'string');
-        this.field = field;
-
-        assert(typeof operator === 'string');
-        this.operator = operator;
-
-        assert(alias === null || typeof alias === 'string');
-        this.alias = alias;
-    }
-
-    clone() {
-        return new AggregationTable(
-            this.table.clone(),
-            this.field,
-            this.operator,
-            this.alias,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Aggregation = AggregationTable;
-Table.Aggregation.prototype.isAggregation = true;
-class SortedTable extends Table {
-    constructor(table, field, direction, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(typeof field === 'string');
-        this.field = field;
-
-        assert(direction === 'asc' || direction === 'desc');
-        this.direction = direction;
-    }
-
-    clone() {
-        return new SortedTable(
-            this.table.clone(),
-            this.field,
-            this.direction,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Sort = SortedTable;
-Table.Sort.prototype.isSort = true;
-class IndexTable extends Table {
-    constructor(table, indices, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(Array.isArray(indices));
-        this.indices = indices;
-    }
-
-    clone() {
-        return new IndexTable(
-            this.table.clone(),
-            this.indices.map((i) => i.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Index = IndexTable;
-Table.Index.prototype.isIndex = true;
-class SlicedTable extends Table {
-    constructor(table, base, limit, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(base instanceof Value);
-        this.base = base;
-
-        assert(limit instanceof Value);
-        this.limit = limit;
-    }
-
-    clone() {
-        return new SlicedTable(
-            this.table.clone(),
-            this.base.clone(),
-            this.limit.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Slice = SlicedTable;
-Table.Slice.prototype.isSlice = true;
-class TableJoin extends Table {
-    constructor(lhs, rhs, in_params, schema) {
-        super(schema);
-
-        assert(lhs instanceof Table);
-        this.lhs = lhs;
-
-        assert(rhs instanceof Table);
-        this.rhs = rhs;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-    }
-
-    clone() {
-        return new TableJoin(
-            this.lhs.clone(),
-            this.rhs.clone(),
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Join = TableJoin;
-Table.Join.prototype.isJoin = true;
-class Window extends Table {
-    constructor(base, delta, stream, schema) {
-        super(schema);
-
-        assert(base instanceof Value); // : Number
-        this.base = base;
-
-        assert(delta instanceof Value); // : Number
-        this.delta = delta;
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-    }
-
-    clone() {
-        return new Window(
-            this.base.clone(),
-            this.delta.clone(),
-            this.stream.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Window = Window;
-Table.Window.prototype.isWindow = true;
-class TimeSeries extends Table {
-    constructor(base, delta, stream, schema) {
-        super(schema);
-
-        assert(base instanceof Value); // : Date
-        this.base = base;
-
-        assert(delta instanceof Value); // : Measure(ms)
-        this.delta = delta;
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-    }
-
-    clone() {
-        return new TimeSeries(
-            this.base.clone(),
-            this.delta.clone(),
-            this.stream.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.TimeSeries = TimeSeries;
-Table.TimeSeries.prototype.isTimeSeries = true;
-class Sequence extends Table {
-    constructor(base, delta, table, schema) {
-        super(schema);
-
-        assert(base instanceof Value); // : Number
-        this.base = base;
-
-        assert(delta instanceof Value); // : Number
-        this.delta = delta;
-
-        assert(table instanceof Table);
-        this.table = table;
-    }
-
-    clone() {
-        return new Sequence(
-            this.base.clone(),
-            this.delta.clone(),
-            this.table.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.Sequence = Sequence;
-Table.Sequence.prototype.isSequence = true;
-class History extends Table {
-    constructor(base, delta, table, schema) {
-        super(schema);
-
-        assert(base instanceof Value); // : Date
-        this.base = base;
-
-        assert(delta instanceof Value); // : Measure(ms)
-        this.delta = delta;
-
-        assert(table instanceof Table);
-        this.table = table;
-    }
-
-    clone() {
-        return new History(
-            this.base.clone(),
-            this.delta.clone(),
-            this.table.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Table.History = History;
-Table.History.prototype.isHistory = true;
-module.exports.Table = Table;
-
-
-/**
- * The base class of all ThingTalk stream expressions.
- *
- * @alias Ast.Stream
- * @extends Ast.Base
- * @abstract
- * @param {Ast.ExpressionSignature|null} schema - type signature of the invoked function
- */
-class Stream extends Base {
-    constructor(schema) {
-        super();
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-}
-Stream.prototype.isStream = true;
-class VarRefStream extends Stream {
-    constructor(name, in_params, schema) {
-        super(schema);
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-    }
-
-    clone() {
-        return new VarRefStream(
-            this.name,
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.VarRef = VarRefStream;
-Stream.VarRef.prototype.isVarRef = true;
-class Timer extends Stream {
-    constructor(base, interval, frequency, schema) {
-        super(schema);
-
-        assert(base instanceof Value);
-        this.base = base;
-
-        assert(interval instanceof Value);
-        this.interval = interval;
-
-        assert(frequency === null || frequency instanceof Value);
-        this.frequency = frequency;
-    }
-
-    clone() {
-        return new Timer(
-            this.base.clone(),
-            this.interval.clone(),
-            this.frequency ? this.frequency.clone() : null,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Timer = Timer;
-Stream.Timer.prototype.isTimer = true;
-class AtTimer extends Stream {
-    constructor(time, expiration_date, schema) {
-        super(schema);
-
-        assert(Array.isArray(time));
-        this.time = time;
-
-        assert(expiration_date === null || expiration_date instanceof Value);
-        this.expiration_date = expiration_date;
-    }
-
-    clone() {
-        return new AtTimer(
-            this.time.map((t) => t.clone()),
-            this.expiration_date ? this.expiration_date.clone() : null,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.AtTimer = AtTimer;
-Stream.AtTimer.prototype.isAtTimer = true;
-class Monitor extends Stream {
-    constructor(table, args, schema) {
-        super(schema);
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(args === null || Array.isArray(args));
-        this.args = args;
-    }
-
-    clone() {
-        return new Monitor(
-            this.table.clone(),
-            this.args ? this.args.map((a) => a) : null,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Monitor = Monitor;
-Stream.Monitor.prototype.isMonitor = true;
-class EdgeNew extends Stream {
-    constructor(stream, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-    }
-
-    clone() {
-        return new EdgeNew(
-            this.stream.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.EdgeNew = EdgeNew;
-Stream.EdgeNew.prototype.isEdgeNew = true;
-class EdgeFilter extends Stream {
-    constructor(stream, filter, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(filter instanceof BooleanExpression);
-        this.filter = filter;
-    }
-
-    clone() {
-        return new EdgeFilter(
-            this.stream.clone(),
-            this.filter.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.EdgeFilter = EdgeFilter;
-Stream.EdgeFilter.prototype.isEdgeFilter = true;
-class FilteredStream extends Stream {
-    constructor(stream, filter, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(filter instanceof BooleanExpression);
-        this.filter = filter;
-    }
-
-    clone() {
-        return new FilteredStream(
-            this.stream.clone(),
-            this.filter.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Filter = FilteredStream;
-Stream.Filter.prototype.isFilter = true;
-class StreamProjection extends Stream {
-    constructor(stream, args, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(Array.isArray(args));
-        this.args = args;
-    }
-
-    clone() {
-        return new StreamProjection(
-            this.stream.clone(),
-            this.args.map((a) => a),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Projection = StreamProjection;
-Stream.Projection.prototype.isProjection = true;
-class ComputeStream extends Stream {
-    constructor(stream, expression, alias, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(expression instanceof ScalarExpression);
-        this.expression = expression;
-
-        assert(alias === null || typeof alias === 'string');
-        this.alias = alias;
-    }
-
-    clone() {
-        return new ComputeStream(
-            this.stream.clone(),
-            this.expression.clone(),
-            this.alias,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Compute = ComputeStream;
-Stream.Compute.prototype.isCompute = true;
-class AliasStream extends Stream {
-    constructor(stream, name, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(typeof name === 'string');
-        this.name = name;
-    }
-
-    clone() {
-        return new AliasStream(
-            this.stream.clone(),
-            this.name,
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Alias = AliasStream;
-Stream.Alias.prototype.isAlias = true;
-class StreamJoin extends Stream {
-    constructor(stream, table, in_params, schema) {
-        super(schema);
-
-        assert(stream instanceof Stream);
-        this.stream = stream;
-
-        assert(table instanceof Table);
-        this.table = table;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-    }
-
-    clone() {
-        return new StreamJoin(
-            this.stream.clone(),
-            this.table.clone(),
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Stream.Join = StreamJoin;
-Stream.Join.prototype.isJoin = true;
-module.exports.Stream = Stream;
-
-
-/**
- * Base class for all expressions that invoke an action.
- *
- * @alias Ast.Action
- * @abstract
- * @param {Ast.ExpressionSignature|null} schema - type signature of this action
- * @property {boolean} isAction - true
- * @property {boolean} isVarRef - true if this is an instance of {@link Ast.Action.VarRef}
- * @property {boolean} isInvocation - true if this is an instance of {@link Ast.Action.Invocation}
- */
-class Action extends Base {
-    constructor(schema) {
-        super();
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-}
-Action.prototype.isAction = true;
-/**
- * An invocation of a locally defined action (i.e. one defined with
- * a `let` statement).
- *
- * @alias Ast.Action.VarRef
- * @extends Ast.Action
- * @param {string} name - the name of the action to invoke
- * @param {Ast.InputParam[]} in_params - the input parameters to pass
- */
-class VarRefAction extends Action {
-    constructor(name, in_params, schema) {
-        super(schema);
-
-        assert(typeof name === 'string');
-        this.name = name;
-
-        assert(Array.isArray(in_params));
-        this.in_params = in_params;
-    }
-
-    clone() {
-        return new VarRefAction(
-            this.name,
-            this.in_params.map((p) => p.clone()),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-
-    toString() {
-        return `VarRef(${this.name}, ${this.in_params.toString()}, )`;
-    }
-}
-Action.VarRef = VarRefAction;
-Action.VarRef.prototype.isVarRef = true;
-/**
- * An invocation of an action in Thingpedia.
- *
- * @alias Ast.Action.Invocation
- * @extends Ast.Action
- * @param {Ast.Invocation} invocation - the function invocation
- * @param {Ast.ExpressionSignature|null} schema - type signature of this action
- */
-class ActionInvocation extends Action {
-    constructor(invocation, schema) {
-        super(schema);
-
-        assert(invocation instanceof Invocation);
-        this.invocation = invocation;
-    }
-
-    clone() {
-        return new ActionInvocation(
-            this.invocation.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-Action.Invocation = ActionInvocation;
-Action.Invocation.prototype.isInvocation = true;
-module.exports.Action = Action;
-
-
-class PermissionFunction extends Base {}
-PermissionFunction.prototype.isPermissionFunction = true;
-class SpecifiedPermissionFunction extends PermissionFunction {
-    constructor(kind, channel, filter, schema) {
-        super();
-
-        assert(typeof kind === 'string');
-        this.kind = kind;
-
-        assert(typeof channel === 'string');
-        this.channel = channel;
-
-        assert(filter instanceof BooleanExpression);
-        this.filter = filter;
-
-        assert(schema === null || schema instanceof ExpressionSignature);
-        this.schema = schema;
-    }
-
-    clone() {
-        return new SpecifiedPermissionFunction(
-            this.kind,
-            this.channel,
-            this.filter.clone(),
-            this.schema ? this.schema.clone() : null
-        );
-    }
-}
-PermissionFunction.Specified = SpecifiedPermissionFunction;
-PermissionFunction.Specified.prototype.isSpecified = true;
-class BuiltinPermissionFunction extends PermissionFunction {
-    clone() {
-        return new BuiltinPermissionFunction();
-    }
-}
-BuiltinPermissionFunction.prototype.isBuiltin = true;
-PermissionFunction.Builtin = new BuiltinPermissionFunction();
-class ClassStarPermissionFunction extends PermissionFunction {
-    constructor(kind) {
-        super();
-
-        assert(typeof kind === 'string');
-        this.kind = kind;
-    }
-
-    clone() {
-        return new ClassStarPermissionFunction(this.kind);
-    }
-}
-PermissionFunction.ClassStar = ClassStarPermissionFunction;
-PermissionFunction.ClassStar.prototype.isClassStar = true;
-class StarPermissionFunction extends PermissionFunction {
-    clone() {
-        return new StarPermissionFunction();
-    }
-}
-StarPermissionFunction.prototype.isStar = true;
-PermissionFunction.Star = new StarPermissionFunction();
-module.exports.PermissionFunction = PermissionFunction;
-
 
 /**
  * The base class of all AST nodes that represent complete ThingTalk
@@ -2018,7 +670,7 @@ class MixinImportStmt extends ImportStmt {
             this.facets.slice(0),
             this.module,
             this.in_params.map((p) => p.clone())
-        )
+        );
     }
 }
 ImportStmt.Mixin = MixinImportStmt;
